// src/types.ts
export interface CoupleInfo {
  groomName: string;
  brideName: string;
  groomFather: string;
  groomMother: string;
  brideFather: string;
  brideMother: string;
  groomPhone: string;
  bridePhone: string;
  groomFatherPhone: string;
  groomMotherPhone: string;
  brideFatherPhone: string;
  brideMotherPhone: string;
  groomAccount: string;
  groomFatherAccount: string;
  groomMotherAccount: string;
  brideAccount: string;
  brideFatherAccount: string;
  brideMotherAccount: string;
  weddingDate: string;
  weddingLocation: string;
  weddingTime: string;
}